restructure to abstract the backend adapter
