+++ arches_orm/arches_django/__init__.py
@@ -1,0 +1,3 @@
+from .wrapper import ArchesDjangoResourceWrapper as adapter
+
+__all__ = ["adapter"]
