+++ arches_orm/arches_django/datatypes/__init__.py
@@ -1,0 +1,12 @@
+from ._register import get_view_model_for_datatype
+from . import (
+    concepts,
+    semantic,
+    resource_instances,
+    string,
+    user
+)
+
+__all__ = [
+    "get_view_model_for_datatype"
+]
