+++ arches_orm/hooks.py
@@ -1,0 +1,3 @@
+from .wrapper import get_adapter
+
+HOOKS = get_adapter().get_hooks()
