+++ arches_orm/models.py
@@ -1,2 +1,4 @@
 # Dummy module, to house dynamically generated
 # resource models.
+from .wkrm import resource_models
+globals().update(resource_models)
